+++ .eslintrc.js
@@ -14,6 +14,7 @@ module.exports = {
     'no-use-before-define': 'off',
     'func-names': 'off',
     'no-restricted-syntax': 'off',
+    'import/extensions': [0, 'never'],
   },
   parserOptions: {
     parser: '@babel/eslint-parser',