+++ functions/backend/index.js
@@ -9,9 +9,8 @@ import validateFirebaseIdToken from '../util/validateFirebaseToken.js';
 import config from '../config.js';
 
 // Routes
-import accessRoutes from './routes/access.js';
-import userRoutes from './routes/user.js';
 import tokenRoutes from './routes/token.js';
+import accessRequestsRoutes from './routes/accessRequests.js';
 
 const app = express();
 
@@ -20,11 +19,8 @@ app.use(cookieParser());
 app.use(express.json());
 app.use(morgan('combined'));
 
+app.use('/accessRequests', accessRequestsRoutes);
 app.use('/token', validateFirebaseIdToken, tokenRoutes);
-
-app.use('/access', accessRoutes);
-
-app.use('/user', validateFirebaseIdToken, userRoutes);
 
 const internal = functions.runWith(config.runtimeOpts).region(config.region).https.onRequest(app);